Add push automation
