+++ README.md
@@ -3,6 +3,8 @@
 Tools for automating various MHubAI-related docker actions from a local node (and GitHub actions).
 
 
-[Link to the automated building README.md](docker-automation/build_and_push/README.md)
+[Link to the automated building README.md](docker-automation/build/README.md)
 
 [Link to the automated testing README.md](docker-automation/testing/README.md)
+
+[Link to the automated pushing README.md](docker-automation/push/README.md)
