--- docker-automation/build_and_push/__init__.py
@@ -1,1 +1,0 @@
-
